+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Test results returned by Runner.get_all_test_results() could not be deleted
 - Aborted tests were missing in Runner.get_all_test_results() results
 - Runner.get_all_test_results() now aborts unclosed tests [Trello 19](https://trello.com/c/mD7TGdoH)
+### Updated
+- Remove Target.shadow method as inconsistent
 
 ## [5.2.0] - 2022-03-24
 ### Added